feat: 검색 필터 및 결과 테이블 컴포넌트 추가, 상수 파일에 시간 슬롯 정의
