+++ src/types.ts
@@ -13,3 +13,12 @@ export interface Schedule {
   range: number[];
   room?: string;
 }
+
+export interface SearchOption {
+  query?: string;
+  grades: number[];
+  days: string[];
+  times: number[];
+  majors: string[];
+  credits?: number;
+}
